+++ providers/flagd/pkg/provider_test.go
@@ -15,7 +15,7 @@ import (
 
 type TestConstructorArgs struct {
 	name    string
-	port    int32
+	port    uint16
 	host    string
 	service flagd.ServiceType
 	options []flagd.ProviderOption
